+++ src/main/java/code/components/move/MoveComponent.java
@@ -7,38 +7,38 @@ import code.components.ComponentType;
  * Created by theo on 3/06/17.
  */
 public class MoveComponent extends Component {
-    private float maxSpeed;
-    private float acc;
-    private float turnRate;
+    private double maxSpeed;
+    private double acc;
+    private double turnRate;
 
-    public MoveComponent(float maxSpeed, float acc, float turnRate) {
+    public MoveComponent(double maxSpeed, double acc, double turnRate) {
         super(ComponentType.MOVE);
         this.maxSpeed = maxSpeed;
         this.acc = acc;
         this.turnRate = turnRate;
     }
 
-    public float getMaxSpeed() {
+    public double getMaxSpeed() {
         return maxSpeed;
     }
 
-    public void setMaxSpeed(float maxSpeed) {
+    public void setMaxSpeed(double maxSpeed) {
         this.maxSpeed = maxSpeed;
     }
 
-    public float getAcc() {
+    public double getAcc() {
         return acc;
     }
 
-    public void setAcc(float acc) {
+    public void setAcc(double acc) {
         this.acc = acc;
     }
 
-    public float getTurnRate() {
+    public double getTurnRate() {
         return turnRate;
     }
 
-    public void setTurnRate(float turnRate) {
+    public void setTurnRate(double turnRate) {
         this.turnRate = turnRate;
     }
 }
